Adds WorkSpaces component.
This component figures out which workspace to load, based on the 'ws'
param from the url string. It then renders the selected workspace,
passing it information about the current user as a prop and a
neuroglancer component as a child. This allows the plugin to render the
page how it wishes and then place the neuroglancer component where it
wants, within that page.