Atlas: fixes selected annotation highlight.
If the annotations shared the exact same title, then all of them would
be highlighted if one was selected. This change fixes the issue by
checking both the title and the timestamp when highlighting. If we get
two annotations with the exact same millisecond timestamp and title,
then oh well.