Atlas/Clio/ImagePicker: Hack to fix 4nm <-> 8nm
HACK!!! Because the initViewer action makes some assumptions about the
dimensions of the dataset, we have to check for the mb20 dataset and
change the dimensions used. This should ideally be fixed in the
initViewer action or the dimensions should be passed as part of the
dataset object from the clio backend. This will become an issue again
as soon as another dataset needs a set of dimensions that differ from
the default 8nm.