ImageSearch: Fixes model name loss on result load
Once the result of the transfer network had been loaded into the page,
the model used disappeared. This was because the result data was being
overwritten as opposed to merged with the new data. Changing the code to
merge the results fixes the issue and the model name sticks around.