Adds code to fetch logged in user roles.
These can be used to determine which parts of the site the currently
logged in user has access to. They are not meant to be used as a
security layer as they can be easily bypassed by setting a different
role in the redux state. The backend will check the roles on the server
side, so we don't have to worry too much about keeping them secure here.